requirements.txt
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
Deprecated==1.2.14
exceptiongroup==1.2.0
fake-useragent==1.5.1
h11==0.14.0
idna==3.7
mss==9.0.1
nodriver==0.27rc3
outcome==1.3.0.post0
pycparser==2.22
PySocks==1.7.1
requests==2.31.0
selenium==4.19.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
tempmail-python==2.3.3
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
urllib3==2.2.1
websockets==12.0
wrapt==1.16.0
wsproto==1.2.0